[animated-polylines] connected polylines to user input
