+++ src/components/css/BarcodeScanner.css
@@ -18,6 +18,8 @@
 .right-side {
   width: 50%;
   padding-right: 25%;
+  max-height: 70vh;
+  overflow-y: auto;
 }
 
 .list-item {
